fix nits about static_cast and binary writer::write(const char[])

+++ src/apps/replication/meta_server/server_state.cpp
@@ -338,7 +338,7 @@ void server_state::query_configuration_by_index(configuration_query_by_index_req
         if (kv.app_name == request.app_name)
         {
             response.err = ERR_OK;
-            response.app_id = i + 1;
+            response.app_id = static_cast<int>(i) + 1;
             response.partition_count = kv.partition_count;
             app_state& app = kv;
             for (auto& idx : request.partition_indices)
